ml/requirements.txt
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
tensorflow==2.15.0
tensorflowjs==4.17.0
pymongo==4.6.1
flask==3.0.0
flask-cors==4.0.0
matplotlib==3.8.2
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0